Initial commit
The 'MATRIX' program removes the xy term from a conic in the form
of Ax^2 + Bxy + Cy^2 + Dx + Ey + F
It also calculates the degrees of rotation that the original conic
had.

The 'MATRIX' program currently relies on the 'CNICDISP' program to
make terms of the equation pop out depending on which one you are
entering. It will be merged with 'MATRIX' in the future.